requirements.txt
PyMySQL==0.9.2
Django<2.0.0,>=1.11.8
google-api-python-client
djangorestframework==3.5.2
markdown 
django-filter==1.1
django-widget-tweaks